requirements.txt
rasa
mysql-connector-python
transformers
python-dotenv
Flask
requests
flask-cors